Added supervision tree to manage depot processes
